requirements.txt
numpy==1.24.3
pandas==2.0.3
streamlit==1.27.0
requests==2.31.0
scikit-learn==1.3.0
joblib==1.3.2
protobuf==3.20.3
gdown==4.7.1
Pillow==10.0.0
catboost==1.2.2
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
